Not redirecting typo issue(#2) resolved
Input link: https://www.youtube.com/techwithrvr?sub_confirmation=1
Shortern link: www.surls.ga/7217a

But really, this link redirects
https://https//www.youtube.com/techwithrvr?sub_confirmation=1

See at the starting of the above link there is a mistake. So to correct this error, I checked the link to see if it had "http://" or "https://" in it. If the substring does not exist, "http://" will be inserted at the beginning of the link. Also, deleted the 'https://' from line number 20. Since it used to add twice. This has resolved the problem.

I hope you are going to accept this pull request.